initial big refractor, probably broke stuff and forgot stuff but oh well baby steps
